\name{binom-internal}
\alias{var.asymp}
\alias{var.cloglog}
\alias{var.logit}
\alias{var.probit}
\alias{ldbinom}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Internal binom functions}
\description{
  Internal functions for binom package.
}
\usage{
var.asymp(p, n = 1)
var.cloglog(p, n = 1)
var.logit(p, n = 1)
var.probit(p, n = 1)
ldbinom(x, size, prob, log = TRUE)
}
%- maybe also 'usage' for other objects documented here.
\details{
  Internal functions for binom package.
}
\author{Sundar Dorai-Raj (sdorairaj@gmail.com) }

\keyword{internal}
